feat (api): refactor object storage interface
