+++ engine/api/objectstore/types.go
@@ -1,0 +1,6 @@
+package objectstore
+
+type Object interface {
+	GetName() string
+	GetPath() string
+}
